requirements.txt
streamlit==1.55.0
groq==1.1.1
python-dotenv==1.1.0
yfinance==0.2.54